requirements.txt
argon2-cffi==21.1.0
asgiref==3.3.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.1.7
django-bootstrap-v5==1.0.1
django-cors-headers==3.10.0
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==9.7.1
fonttools==4.28.1
idna==3.3
importlib-metadata==2.1.1
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2==2.9.1
pycparser==2.21
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==302
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.26.0
scipy==1.7.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
sympy==1.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tomli==1.2.2
toposort==1.7
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.4.1
